requirements.txt
fastapi>=0.115.10,<0.116.0
uvicorn>=0.34.0,<0.35.0
rembg>=2.0.63,<3.0.0
pillow>=11.1.0,<12.0.0
python-multipart>=0.0.20,<0.0.21
pydantic>=2.10.6,<3.0.0
onnxruntime>=1.17.3
pymongo>=4.6.1
python-dotenv>=1.0.0
cloudinary>=1.33.0